Last update. Had heat index logic missing from Weather.isGoodWeather()
This is ready to turn in. A summary of changes and lessons learned.

The logic is from a top down class design. Starting with class Event, the basic data fields and methods are added to be inherited. This includes the constructor, which assigns all arguments that are shared by subclasses. The format date helper method, which formats the dateTime object into a simpler date and time format. As well as the first toString() method, which returns a string with the information shared by all subclasses.
Each subclass builds on this. The Constructors first pass arguments above to the parent constructor until stopping at Event.java. This Constructor chaining cuts down on code by reusing the constructors of each level in the hierarchy. The toString() method does the same by return String.format() at each sublevel, passing super.toString() as an argument in a subclass' overriding method. This again cuts down on code by reusing the parent's toString() method code.

**Lessons Learned**
+Proper Class Design before coding will save so much time. I had to rewrite a lot of code because it was redundant as i progressed through the project.
+Method Overriding and Polymorphism makes life so much easier. The toString() method could have been much more incontinent if I had not used the super.toString chaining up the hierarchy to prevent duplicate code.
+Constructor chaining is more confusing than I first thought. in order to chain the object counter for each class type, I had to duplicate to the code on each class. I am sure there is a better way to do this. However the Test class was not allowed to be touched, so I had to do this. If I could of changed the Test class, I might of centralized the method in Event.java and had the calling class act as a sort of argument. I need to research if this is possible.

Overall, I learned a lot and was able to properly understand how inheritance works.